requirements.txt
streamlit
streamlit-drawable-canvas
fpdf2
Pillow
pandas
